requirements.txt
google-genai>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
